pluginfs: add alloc_inode and destroy_inode operations
As pointed out by Sergey we cannot block/schedule in plgfs_inode_set, because
this callback is called with inode_hash_lock spinlock held. Meaning the
allocation of inode_info here with GFP_KERNEL is wrong. To fix this the
inode_info allocation and destruction is moved to alloc_inode and destroy_inode
callbacks.

The original Sergey's post about this problem could be found here
http://pluginfs.org/pipermail/devel/2014-January/000008.html

Signed-off-by: Frantisek Hrbata <fhrbata@pluginfs.org>
Tested-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>